+++ agent/src/main/java/com/teamscale/jacoco/agent/commandline/ICommand.java
@@ -5,6 +5,10 @@
 +-------------------------------------------------------------------------*/
 package com.teamscale.jacoco.agent.commandline;
 
+import com.teamscale.jacoco.agent.options.AgentOptionParseException;
+
+import java.io.IOException;
+
 /**
  * Interface for commands: argument parsing and execution.
  */
@@ -14,7 +18,7 @@ public interface ICommand {
 	 * Makes sure the arguments are valid. Must return all detected problems in the
 	 * form of a user-visible message.
 	 */
-	Validator validate();
+	Validator validate() throws AgentOptionParseException, IOException;
 
 	/**
 	 * Runs the implementation of the command. May throw an exception to indicate